+++ R/utils-add_p_tests.R
@@ -4,6 +4,7 @@
 
 add_p_test_t.test <- function(data, variable, by, test.args, conf.level = 0.95, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars"), ...)
 
   rlang::inject(
     cardx::ard_stats_t_test(
@@ -18,6 +19,7 @@ add_p_test_t.test <- function(data, variable, by, test.args, conf.level = 0.95, 
 
 add_p_test_wilcox.test <- function(data, variable, by, test.args, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars"), ...)
 
   rlang::inject(
     cardx::ard_stats_wilcox_test(
@@ -38,6 +40,8 @@ add_p_test_wilcox.test <- function(data, variable, by, test.args, ...) {
 
 add_p_test_mcnemar.test <- function(data, variable, by, group, test.args, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("adj.vars"), ...)
+  check_length(group, 1L)
   warn_unbalanced_pairs(data, by, variable, group)
 
   rlang::inject(
@@ -53,6 +57,7 @@ add_p_test_mcnemar.test <- function(data, variable, by, group, test.args, ...) {
 
 add_p_test_mcnemar.test_wide <- function(data, variable, by, test.args, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars"), ...)
 
   rlang::inject(
     cardx::ard_stats_mcnemar_test(
@@ -66,6 +71,7 @@ add_p_test_mcnemar.test_wide <- function(data, variable, by, test.args, ...) {
 
 add_p_test_chisq.test <- function(data, variable, by, test.args, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars"), ...)
 
   rlang::inject(
     cardx::ard_stats_chisq_test(
@@ -91,6 +97,7 @@ add_p_test_chisq.test.no.correct <- function(data, variable, by, test.args, ...)
 
 add_p_test_mood.test <- function(data, variable, by, test.args, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars"), ...)
 
   rlang::inject(
     cardx::ard_stats_mood_test(
@@ -105,6 +112,7 @@ add_p_test_mood.test <- function(data, variable, by, test.args, ...) {
 
 add_p_test_kruskal.test <- function(data, variable, by, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars", "test.args"), ...)
 
   cardx::ard_stats_kruskal_test(
     data = data,
@@ -115,6 +123,7 @@ add_p_test_kruskal.test <- function(data, variable, by, ...) {
 
 add_p_test_fisher.test <- function(data, variable, by, test.args, conf.level = 0.95, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars"), ...)
 
   rlang::inject(
     cardx::ard_stats_fisher_test(
@@ -147,6 +156,7 @@ add_p_test_aov <- function(data, variable, by, ...) {
 
 add_p_test_oneway.test <- function(data, variable, by, test.args, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars"), ...)
 
   rlang::inject(
     cardx::ard_stats_oneway_test(
@@ -159,6 +169,7 @@ add_p_test_oneway.test <- function(data, variable, by, test.args, ...) {
 
 add_p_test_mood.test <- function(data, variable, by, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars", "test.args"), ...)
 
   rlang::inject(
     cardx::ard_stats_mood_test(
@@ -172,10 +183,12 @@ add_p_test_mood.test <- function(data, variable, by, ...) {
 add_p_test_lme4 <- function(data, variable, by, group, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
   check_pkg_installed("lme4", reference_pkg = "cardx")
+  check_null(c("test.args", "adj.vars"), ...)
 
   if (is_empty(group)) {
     cli::cli_abort("The {.arg group} argument cannot be missing for {.val lme4} tests.")
   }
+  check_length(group, 1L)
 
   cardx::ard_stats_anova(
     x = data |> tidyr::drop_na(any_of(c(variable, by, group))),
@@ -201,6 +214,7 @@ add_p_tbl_summary_paired.t.test <- function(data, variable, by, group, test.args
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
   check_length(group, 1L)
   warn_unbalanced_pairs(data, by, variable, group)
+  check_null(c("adj.vars"), ...)
 
   rlang::inject(
     cardx::ard_stats_paired_t_test(
@@ -216,6 +230,7 @@ add_p_tbl_summary_paired.t.test <- function(data, variable, by, group, test.args
 
 add_p_tbl_summary_paired.wilcox.test <- function(data, variable, by, group, test.args, conf.level = 0.95, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("adj.vars"), ...)
   check_length(group, 1L)
   warn_unbalanced_pairs(data, by, variable, group)
 
@@ -231,8 +246,9 @@ add_p_tbl_summary_paired.wilcox.test <- function(data, variable, by, group, test
   )
 }
 
-add_p_test_prop.test <- function(data, variable, by, group, test.args, conf.level = 0.95, ...) {
+add_p_test_prop.test <- function(data, variable, by, test.args, conf.level = 0.95, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("adj.vars", "group"), ...)
 
   rlang::inject(
     cardx::ard_stats_prop_test(
@@ -247,6 +263,16 @@ add_p_test_prop.test <- function(data, variable, by, group, test.args, conf.leve
 
 add_p_test_ancova <- function(data, variable, by, adj.vars, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "test.args"), ...)
+
+  if (dplyr::n_distinct(data[[by]], na.rm = TRUE) != 2L) {
+    cli::cli_abort(
+      c("Error in {.val ancova} model for variable {.val {variable}}.",
+        i = "This inplementation requires the {.val by} variable to have {.val {2}} levels,
+             but it has {.val {dplyr::n_distinct(data[[by]], na.rm = TRUE)}}"),
+      call = get_cli_abort_call()
+    )
+  }
 
   cardx::ard_regression_basic(
     x =
@@ -259,7 +285,7 @@ add_p_test_ancova <- function(data, variable, by, adj.vars, ...) {
     dplyr::filter(
       .data$variable %in% .env$by,
       .data$stat_name %in% c("estimate", "p.value", "statistic", "std.error", "conf.low", "conf.high")
-    ) |>
+    ) %>%
     dplyr::bind_rows(
       .,
       dplyr::filter(., dplyr::row_number() == 1L) |>
@@ -279,6 +305,7 @@ add_p_test_ancova <- function(data, variable, by, adj.vars, ...) {
 
 add_p_test_cohens_d <- function(data, variable, by, test.args, conf.level = 0.95, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars"), ...)
 
   rlang::inject(
     cardx::ard_effectsize_cohens_d(
@@ -294,6 +321,7 @@ add_p_test_cohens_d <- function(data, variable, by, test.args, conf.level = 0.95
 # TODO: Add this function to the test xlsx file
 add_p_test_paired_cohens_d <- function(data, variable, by, test.args, group, conf.level = 0.95, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("adj.vars"), ...)
   check_length(group, 1L)
   warn_unbalanced_pairs(data, by, variable, group)
 
@@ -312,6 +340,7 @@ add_p_test_paired_cohens_d <- function(data, variable, by, test.args, group, con
 # TODO: Add this function to the test xlsx file
 add_p_test_hedges_g <- function(data, variable, by, test.args, conf.level = 0.95, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars"), ...)
 
   rlang::inject(
     cardx::ard_effectsize_hedges_g(
@@ -327,6 +356,7 @@ add_p_test_hedges_g <- function(data, variable, by, test.args, conf.level = 0.95
 # TODO: Add this function to the test xlsx file
 add_p_test_paired_hedges_g <- function(data, variable, by, test.args, group, conf.level = 0.95, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("adj.vars"), ...)
   check_length(group, 1L)
   warn_unbalanced_pairs(data, by, variable, group)
 
@@ -344,6 +374,7 @@ add_p_test_paired_hedges_g <- function(data, variable, by, test.args, group, con
 
 add_p_test_smd <- function(data, variable, by, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("group", "adj.vars", "test.args"), ...)
 
   cardx::ard_smd_smd(
     data = data,
@@ -355,6 +386,7 @@ add_p_test_smd <- function(data, variable, by, ...) {
 add_p_test_emmeans <- function(data, variable, by, adj.vars, conf.level = 0.95,
                                type, group = NULL, ...) {
   check_pkg_installed("cardx", reference_pkg = "gtsummary")
+  check_null(c("test.args"), ...)
 
   if (!is.null(group)) check_pkg_installed("lme4", reference_pkg = "cardx")
   if (inherits(data, "survey.design")) check_pkg_installed("survey", reference_pkg = "cardx")
@@ -419,6 +451,7 @@ add_p_test_emmeans <- function(data, variable, by, adj.vars, conf.level = 0.95,
   )
 }
 
+# UTILITY FUNCTIONS ------------------------------------------------------------
 .extract_data_frame <- function(x) {
   if (is.data.frame(x)) {
     return(x)
@@ -447,5 +480,23 @@ warn_unbalanced_pairs <- function(data, by, variable, group) {
        from the comparison due to unbalanced missingness within group."
     )
   }
+  invisible()
+}
+
+check_null <- function(x, variable = get("variable", envir = caller_env()),
+                       call = get_cli_abort_call(), ...) {
+  dots <- dots_list(...)
+  walk(
+    x,
+    \(.x) {
+      if (!is_empty(dots[[.x]])) {
+        cli::cli_abort(
+          message = "The {.arg {.x}} argument is not used for the test selected for of variable {.val {variable}}.",
+          call = call
+        )
+      }
+    }
+  )
+
   invisible()
 }